Fixes #46
